requirements.txt
PyHamcrest==2.0.2
pytest==6.2.2
requests==2.25.1
selenium==3.141.0
urllib3==1.26.4
